docs: guides overhaul
This PR is a complete overhaul of the `guides/` directory. While trying to wrap my head around the library, I found the existing Guides to be a bit confusing and lacking.

Using the existing Guides as a starting point, I have re-structured and added a lot of new content to make the Guides more useful and easier to navigate. I also paid attention to the style and formatting to make it easier to read and understand, while sticking to the existing style as much as possible.

This is _my_ understanding of the library, which I pieced together from the existing Guides, examples, API Docs and delving into the `src/` and `development/` source code. Props to TypeScript for helping out with the latter.

I am sure there are some errors and omissions, so feedback welcome and I am happy to make any changes as needed.

## Changes

In this new form, the `guides/` directory contains the following Markdown pages:

1. Getting Started
2. Adapters
3. Modes
4. Styling
5. Events
6. Development

### Restructuring / Navigation

- All pages in the `guides/` directory now have a numeric prefix to control the order in which they appear. The exception is the `CODE_OF_CONDUCT.md` file, which I have left as-is.
- At the end of each page there is a list of links to the other pages in the `guides/` directory, so it is easy to navigate between them.

### 1.Getting Started

- Added additional Key Concepts content:
  - Store
  - Adapters
  - Modes
- Improved Installation instructions.

### 2.Adapters

Added sections:

- Available Adapters
- Using Adapters
- Examples:
  - Leaflet
  - MapLibre
  - Google Maps
  - MapBox

### 3.Modes

Added sections:

- Mode Names
- Drawing Modes
- Selection Mode
- Render Mode
- Adding Modes
- Enabling Modes
- Loading Features

### 4.Styling

Added sections:

- Drawing Modes:
  - Points
  - Lines
  - Polygons
  - Freehand
  - Great Circle
  - Circle
  - Rectangle
- Selection Mode
  - Points
  - Lines
  - Polygons
  - Selection Points
  - Mid Points
- Render Mode
  - Points
  - Lines
  - Polygons

### 5.Events

Restructured sections into:

- Native Events
- Terra Draw Events

### 6.Development

Mostly minor tweaks and improvements.

### General / Other

- Added inline Alerts to highlight important information.
- Added lang to all code blocks.
- Lots of minor tweaks to improve formatting and readability.

### Removed

- `guides/COMMON_PATTERNS.md` - content broken up and moved to other pages.
- `guides/CONTRIBUTING.md` - content moved to `DEVELOPMENT.md#contributing`.

The README.md file in the root directory, as well as the `docs/index.html` has been updated to reflect these Guide location changes.